feat : main, sub 컴포넌트 분리 / 기간설정 가능하게 변경(#10)
